# [bencode] Rewrite nrepl.bencode