I freshened up the implementation as discussed in #322 and made performance improvements where it made sense to do so. Although the spirit of the implementation remained mostly the same, there are a few significant changes:

- `read-token` is gone, reading dispatch happens within the `read-bencode` function for better efficiency.
- Rewrote writing dispatch from multimethods to protocols. Introduced a separate private protocol function `write-bencode*`, the public `write-bencode` is a public function with old argument order.

I've generated a stress test data structure (~10Mb when serialized as EDN) to benchmark. Here are the results:

##### Encoding
`Before: Time per call: 230.37 ms   Alloc per call: 293,593,739b`
`After:  Time per call: 89.59 ms    Alloc per call: 114,506,367b`
Improvement: 2.5x in time and 2.5x in allocations.

##### Decoding
`Before: Time per call: 263.36 ms   Alloc per call: 374,434,013b`
`After:  Time per call: 49.39 ms    Alloc per call: 76,781,066b`
Improvement: 5.3x in time and 4.9x in allocations.

Overall, I don't think this would influence the UX of e.g. CIDER much because the data structures that go through cider-nrepl are usually small, and the bottleneck would most likely be on the Emacs side, not on the Clojure side. Anyway, a win is a win.