@bbatsov 2024-05-22
Impressive performance gains! Great work, Sashko! 🚀 

At at a glance I don't see anything concerning in the code changes that you made, so I'm pondering now whether to ship this with nREPL 1.1.2 or nREPL 1.2. I'm leaning slightly towards 1.2, together with the new agent. Any preferences regarding this?