@bbatsov 2024-05-22
Seems all you need to do now is make `cljfmt` happy (it complains that something isn't indented properly) and we can merge this PR and cut 1.2.0-beta1.