@alexander-yakushev 2024-05-22
All good now. Also updated the changelog mentioning the jvmti change which I forgot.