@alexander-yakushev 2024-05-22
Not good, one test failed with OOM. Reduced num-tests for generative testing.