@bbatsov 2024-05-22
Seems there's still some OOM flakiness https://app.circleci.com/pipelines/github/nrepl/nrepl/556/workflows/966d5bbd-ba07-4b41-9eb4-d99e8a034677/jobs/9124