@alexander-yakushev 2024-06-11
Hm, I didn't consider that somebody may be relying on this; after all, the number of types that Bencode declares is a closed set. You can extend `BencodeSerializable` protocol for your types; or we can consider reverting this back to multimethod dispatch. @bbatsov thoughts?