@bbatsov 2024-06-11
> Hm, I didn't consider that somebody may be relying on this; after all, the number of types that Bencode declares is a closed set. 

Same.

> You can either extend BencodeSerializable protocol for your types;

I'd prefer this direction.