@bbatsov 2024-06-11
> (for software extending nrepl.bencode/write-bencode by adding methods like is the case for FlowStorm nrepl middleware)

@jpmonettas Btw, out of curiosity why were you doing something like this?