@jpmonettas 2024-06-11
@alexander-yakushev it is ok as it is, I just figured that I can fix it on my side, not relying on that, so I will not be pushing for any revert.

@bbatsov 
> out of curiosity why were you doing something like this?

So, different ops on my middleware were dealing with maps with keys pointing to custom types. So instead of doing updates on each piece of data before sending the response I went and used write-bencode extension as a serialization layer. Maybe it was hacky.

Anyway I figured out I can do it without it, by converting the data before sending the response, there aren't many cases, so no worries.