@alexander-yakushev 2024-06-11
I think it's not really different from a regular public function in a "non-public" namespace that is renamed or deleted or has arglist changed or even retains all visual compatibility but suddenly starts behaving differently. I personally don't agree that any public function is automatically a public API. This kind of thinking leads us to Java land where everything is paranoidally private.