@jpmonettas 2024-06-11
Yeah, I don't agree either that every non private function is automatically considered part of a public api, but since multimethods and protocols are normally used because of their open extension capabilities, they lead me to think that the library authors added them as extension points for everybody, not just for them. Like in this case, write-bencode being a multimethod looked to me like it was designed for people to write anything into bencode.