@alexander-yakushev 2024-06-11
Maybe, this was a case of incomplete documentation. In any case, I don't know what the original author had in mind in 2012, but in the current implementation the protocol is used purely as a developer's convenience.