@bbatsov 2024-06-11
> Maybe, this was a case of incomplete documentation. In any case, I don't know what the original author had in mind in 2012, but in the current implementation the protocol is used purely as a developer's convenience.

Pretty sure this was the case back in the day as well - I've often used multimethods in the past just because I thought the code was more elegant this way. (without actually anticipating someone would add more down the road)

Anyways, it's always great to learn how someone is your using your libraries - especially if that's something you didn't think about. :-) Sorry about any issues we might have caused!